requirements.txt
flask==3.0.0
Pillow==10.4.0
requests==2.31.0 